backend/requirements.txt
# Core AI/ML Libraries - Updated to compatible versions
langchain>=0.3.0,<0.4.0
langchain-core>=1.0.0,<2.0.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langchain-groq>=0.1.0
langchain-community>=0.3.0
langgraph>=0.2.0,<0.3.0
langsmith>=0.2.0,<1.0.0
openai>=1.12.0  # Only for embeddings (optional if using sentence-transformers)
anthropic>=0.18.1
groq>=0.4.2  # Groq SDK for LLM inference
sentence-transformers>=2.2.2  # Free local embeddings alternative
moondream>=0.1.0  # Moondream AI for vision-language tasks (floor plan analysis)

# Vector Database
qdrant-client==1.7.0
pinecone-client==3.0.0

# Computer Vision
opencv-python==4.9.0.80
ultralytics==8.1.0  # YOLOv8
torch==2.2.0
torchvision==0.17.0
pillow==10.2.0
transformers==4.37.2

# PDF Processing
pypdf2==3.0.1
pdfplumber==0.10.3
pymupdf==1.23.8
pdf2image==1.16.3

# OCR
pytesseract==0.3.10
easyocr==1.7.0

# API & Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Async & Task Queue
celery==5.3.4
redis==5.0.1
aiofiles==23.2.1

# Utilities
python-dotenv==1.0.0
numpy==1.26.3
pandas==2.1.4
openpyxl==3.1.2  # Excel export
python-multipart==0.0.6
httpx==0.26.0

# Evaluation & Monitoring
wandb==0.16.0
tensorboard==2.15.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
